Changed: #825 Remove all warning when compiling Ryzom on Linux

+++ code/nel/include/nel/3d/cluster.h
@@ -111,7 +111,7 @@ public:
 	void link (CPortal* portal);
 	void unlink (CPortal* portal);
 
-	uint32 getNbPortals() {return _Portals.size();}
+	uint32 getNbPortals() {return (uint32)_Portals.size();}
 	CPortal* getPortal(uint32 pos) {return _Portals[pos];}
 
 	/// Serial
