+++ code/nel/include/nel/3d/ps_mesh.h
@@ -266,7 +266,7 @@ public:
 	{
 		min = _MinAngularVelocity;
 		max = _MaxAngularVelocity;
-		return _PrecompBasis.size();
+		return (uint32)_PrecompBasis.size();
 	}
